feat: Add context auth integration
