+++ frontend/app/(auth)/login.tsx
@@ -2,6 +2,8 @@ import { Pressable, StyleSheet, Text, View } from 'react-native'
 import React, { useState } from 'react'
 import Icon from 'react-native-vector-icons/Ionicons';
 import FontAwesome from 'react-native-vector-icons/FontAwesome5'
+import { router } from 'expo-router'
+import { useAuth } from '../contexts/AuthContext'
 
 
 //Themed Components
@@ -10,12 +12,23 @@ import ThemedTextInput from "../../components/ThemedTextInput"
 
 
 const Login = () => {
+  const { signIn } = useAuth()
 
   const [registerNumber, setRegisterNumber] = useState('')
   const [password, setPassword] = useState('')
+  const [submitAttempted, setSubmitAttempted] = useState(false)
 
-  const handleSubmit =() => {
-    console.log("User : ", registerNumber, "Pass : ", password, "Logged In")
+  const handleSubmit = async () => {
+    setSubmitAttempted(true)
+    const usernameEmpty = registerNumber.trim().length === 0
+    const passwordEmpty = password.trim().length === 0
+    if (usernameEmpty || passwordEmpty) return
+
+    // Demo logic: infer role; replace with real auth API
+    const isAdmin = registerNumber.trim().toLowerCase().includes('admin')
+    await signIn({ role: isAdmin ? 'admin' : 'user', registerNumber })
+    // Send to the index route which will redirect based on role
+    router.replace('/')
   }
 
 
@@ -36,9 +49,15 @@ const Login = () => {
     
     <View style={styles.inputContainer}>
       <Icon name='person' size={20} style={[{marginHorizontal: 10}, {color: '#65758c'}]}/>
-      <ThemedTextInput placeholder="Registration Number" placeholderTextColor={'#65758c'} style={{width:"80%"}}
+      <ThemedTextInput placeholder="Username" placeholderTextColor={'#65758c'} style={{width:"80%"}}
       onChangeText={setRegisterNumber} value={registerNumber}/>
     </View>
+    {submitAttempted && registerNumber.trim().length === 0 ? (
+      <View style={styles.errorRow}>
+        <Icon name="information-circle" size={16} color="#9e0202" />
+        <Text style={styles.errorText}>Username is required</Text>
+      </View>
+    ) : null}
 
     <Spacer height={20}/>
 
@@ -47,6 +66,12 @@ const Login = () => {
       <ThemedTextInput placeholder="Password" placeholderTextColor={'#65758c'} style={{width:"80%"}}
       onChangeText={setPassword} value={password} secureTextEntry/>
     </View>
+    {submitAttempted && password.trim().length === 0 ? (
+      <View style={styles.errorRow}>
+        <Icon name="information-circle" size={16} color="#9e0202" />
+        <Text style={styles.errorText}>Password is required</Text>
+      </View>
+    ) : null}
 
     <Spacer height={40}/>
 
@@ -69,6 +94,18 @@ const styles = StyleSheet.create({
         flex: 1,
         justifyContent: 'center',
         alignItems: 'center',
+    },
+    errorRow: {
+      flexDirection: 'row',
+      alignItems: 'center',
+      width: '80%',
+      marginTop: 6,
+      gap: 6,
+    },
+    errorText: {
+      color: '#9e0202',
+      fontSize: 12,
+      fontWeight: '600',
     },
 
     inputContainer:{
