+++ frontend/app/(dashboard-admin)/adminCreateEvent.tsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const CreateEvent = () => {
+  return (
+    <View>
+      <Text>CreateEvent</Text>
+    </View>
+  )
+}
+
+export default CreateEvent
+
+const styles = StyleSheet.create({})
