+++ frontend/app/(dashboard-admin)/adminHome.tsx
@@ -1,0 +1,18 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const AdminHome = () => {
+  return (
+    <View>
+      <Text>AdminHome</Text>
+    </View>
+  )
+}
+
+export default AdminHome
+
+const styles = StyleSheet.create({
+    container:{
+        flex: 1,
+    }
+})
