+++ frontend/app/(dashboard-admin)/adminSearch.tsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const AdminSearch = () => {
+  return (
+    <View>
+      <Text>AdminSearch</Text>
+    </View>
+  )
+}
+
+export default AdminSearch
+
+const styles = StyleSheet.create({})
